+++ image_gen/image/__init__.py
@@ -1,3 +1,0 @@
-from image_gen.image.generator import ImageGenerator
-
-__all__ = ["ImageGenerator"]