+++ media_utils/image/__init__.py
@@ -1,0 +1,3 @@
+from media_utils.image.generator import ImageGenerator
+
+__all__ = ["ImageGenerator"]